+++ content/en-US.json
@@ -84,5 +84,15 @@
   },
   "openAccount": {
     "title": "Open Account"
+  },
+  "Compliance": {
+    "title": "Compliance",
+    "caption": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt.",
+    "content": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed ornare blandit libero in fermentum. Integer placerat nisi quis justo interdum, nec consectetur nisl finibus. Vestibulum placerat magna non urna vehicula fringilla. Nullam vel purus ac lorem pulvinar posuere a ut neque. Quisque suscipit fermentum hendrerit. Integer augue nulla, pellentesque ac pellentesque posuere, sollicitudin id odio. Aenean ultrices ex sit amet neque posuere venenatis. Fusce volutpat rhoncus fermentum. Integer at lobortis nisi, id ultricies elit. Donec ut consectetur est, eget mattis orci. Nulla ut porta tellus. Integer non cursus quam. Donec quis placerat eros.\n\nSed maximus augue magna, ultrices elementum nibh efficitur in. Maecenas sed tellus vitae enim accumsan tempus quis quis orci. Cras ac est non mauris convallis faucibus. Praesent ac lectus lorem. Nullam eget interdum ante. Praesent eu purus ac quam porttitor vestibulum. Curabitur accumsan commodo erat, eu facilisis turpis. Praesent eget purus vel mi tincidunt blandit pretium a massa. In hac habitasse platea dictumst. Praesent imperdiet ut ante dignissim cursus. Sed magna mi, consectetur porta ullamcorper nec, maximus nec purus. Praesent facilisis viverra quam, quis porta metus."
+  },
+  "Ouvidoria": {
+    "title": "Ouvidoria",
+    "caption": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt.",
+    "content": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed ornare blandit libero in fermentum. Integer placerat nisi quis justo interdum, nec consectetur nisl finibus. Vestibulum placerat magna non urna vehicula fringilla. Nullam vel purus ac lorem pulvinar posuere a ut neque. Quisque suscipit fermentum hendrerit. Integer augue nulla, pellentesque ac pellentesque posuere, sollicitudin id odio. Aenean ultrices ex sit amet neque posuere venenatis. Fusce volutpat rhoncus fermentum. Integer at lobortis nisi, id ultricies elit. Donec ut consectetur est, eget mattis orci. Nulla ut porta tellus. Integer non cursus quam. Donec quis placerat eros.\n\nSed maximus augue magna, ultrices elementum nibh efficitur in. Maecenas sed tellus vitae enim accumsan tempus quis quis orci. Cras ac est non mauris convallis faucibus. Praesent ac lectus lorem. Nullam eget interdum ante. Praesent eu purus ac quam porttitor vestibulum. Curabitur accumsan commodo erat, eu facilisis turpis. Praesent eget purus vel mi tincidunt blandit pretium a massa. In hac habitasse platea dictumst. Praesent imperdiet ut ante dignissim cursus. Sed magna mi, consectetur porta ullamcorper nec, maximus nec purus. Praesent facilisis viverra quam, quis porta metus."
   }
 }
