+++ public/sitemap.xml
@@ -1,38 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>www.yourwebsite.com</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/cadastro</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/cadastro/Cadastro</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/cambio/corporativo</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/cambio/corporativo/Corporativo</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/cambio/pessoa-fisica</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/cambio/pessoa-fisica/PessoaFisica</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/compliance</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/compliance/Compliance</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/faq</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/faq/FAQ</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/institucional</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/institucional/Institucional</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/ouvidoria</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/ouvidoria/Ouvidoria</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/trabalhe-conosco</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/trabalhe-conosco/TrabalheConosco</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/pt-BR/cadastro</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/en-US/cadastro</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/pt-BR/compliance</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/en-US/compliance</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/pt-BR/faq</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/en-US/faq</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/pt-BR/ouvidoria</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/en-US/ouvidoria</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/pt-BR/trabalhe-conosco</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/en-US/trabalhe-conosco</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/pt-BR/cambio/corporativo</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/en-US/cambio/corporativo</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/pt-BR/cambio/pessoa-fisica</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/en-US/cambio/pessoa-fisica</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/pt-BR</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/en-US</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/pt-BR/institucional</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
-<url><loc>www.yourwebsite.com/en-US/institucional</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-15T01:44:45.708Z</lastmod></url>
+<url><loc>www.yourwebsite.com</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-16T14:51:19.688Z</lastmod></url>
+<url><loc>www.yourwebsite.com/compliance</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-16T14:51:19.688Z</lastmod></url>
+<url><loc>www.yourwebsite.com/institucional</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-16T14:51:19.688Z</lastmod></url>
+<url><loc>www.yourwebsite.com/next/dist/pages/_error</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-16T14:51:19.688Z</lastmod></url>
+<url><loc>www.yourwebsite.com/ouvidoria</loc><changefreq>daily</changefreq><priority>0.7</priority><lastmod>2021-11-16T14:51:19.688Z</lastmod></url>
 </urlset>
